requirements.txt
absl-py==2.1.0
accelerate==1.6.0
addict==2.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
blis==1.2.0
blobfile==3.0.0
block.bootstrap.pytorch==0.1.6
bootstrap.pytorch==0.0.13
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1725267488082/work
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725560520483/work
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.21.0
comm==0.2.2
confection==0.1.5
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
dash==3.0.0
decorator==5.2.1
decord==0.6.0
deepspeed==0.16.7
einops==0.8.1
en_core_web_lg @ file:///data1/sunmingyu/xiaoxia_laboratory/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
exceptiongroup==1.2.2
executing==2.2.0
fairscale==0.4.4
fastjsonschema==2.21.1
filelock==3.16.1
flash-attn @ file:///data1/sunmingyu/xiaoxia_laboratory/my_study/flash_attn-2.7.3%2Bcu11torch2.4cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=329d19f64050d990296a94f22d63990c168a4f658fa28623225f8abf08170380
Flask==3.0.3
fonttools==4.56.0
fsspec==2024.9.0
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1733462536562/work
google-auth==2.34.0
google-auth-oauthlib==1.0.0
grpcio==1.66.1
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
h5py==3.11.0
hjson==3.1.0
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
huggingface-hub==0.30.2
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
imageio==2.37.0
immutabledict==4.2.0
importlib_metadata==8.5.0
ipdb==0.13.13
ipython==8.34.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
lxml==5.3.2
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
msgpack==1.1.0
munch==4.0.0
murmurhash==1.0.12
narwhals==1.31.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.4
nltk==3.9.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1745118901167/work/dist/numpy-2.2.5-cp310-cp310-linux_x86_64.whl#sha256=d8a9c6cae6e78c7a886e39e6ad81ccc3504cc585d1b0944c86fe1308fa99816d
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.19.0
opencv-python==4.11.0.86
ortools==9.11.4210
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.10.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==6.0.1
preshed==3.0.9
pretrainedmodels==0.7.4
prompt_toolkit==3.0.50
protobuf==5.26.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pycryptodomex==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pyquaternion==0.9.9
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1737454647378/work
referencing==0.36.2
regex==2024.9.11
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.23.1
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.0
SharedArray==3.2.4
shellingham==1.5.4
six==1.17.0
skipthoughts==0.0.1
smart-open==7.1.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
stanfordnlp==0.2.0
stringcase==1.2.0
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1736248176451/work
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
thinc==8.3.4
threadpoolctl==3.6.0
tifffile==2025.3.13
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
tomli==2.2.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.51.3
triton==3.0.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==1.17.2
zipp==3.20.2
zstandard==0.23.0
